Fold act duplicate-member tests into ActClaimDeserializationTests as regular tests
Moves the three act-claim round-trip tests (single act member emission, last-wins duplicate
resolution, and no duplicate members from a validated identity reused as an actor) from the separate
ActEdgeCaseTests file into ActClaimDeserializationTests, and deletes the edge-case file. Reframes the
doc comments from 'FINDING #'/'currently FAIL' investigation notes into regular regression-test
descriptions; the fixes for these behaviors already shipped in this branch.

Brings over the tests' private helpers (IssueToken, PayloadOf, ActClaimOf, CountMembers, RawJson,
CreateInboundDelegationToken, ActorIdentifyingClaims, the symmetric-key ValidationParameters) and adds
an ITestOutputHelper constructor for the diagnostic reporting in the no-duplicate-members test.

All ActClaim tests (48) and the full JWT suite (721 passed, 1 skipped) remain green.
